+++ index.php
@@ -7,9 +7,13 @@
     <title>Sudarshan Giri</title>
 </head>
 <body>
-    <p><h2>RESUME REGISTRY OF </h2></p>
+    <p><h2>RESUME REGISTRY OF SUDARSHAN GIRI</h2></p>
     <?php
-   
+    if(isset($_SESSION['message'])){
+        echo $_SESSION['message'];
+
+    }
+    unset($_SESSION['message']);
     if(isset($_SESSION['user_id'])){
         echo " <p><a href='logout.php'> Logout</a></p>";
 
@@ -36,16 +40,24 @@
       $stmt->execute();
       //$result = $stmt->fetchAll(PDO::FETCH_ASSOC);
       ?><table border="1">
-      <tr><th>Name</th><th>Headline</th><th>Action</th><tr>
+      <tr><th>Name</th><th>Headline</th>                    <?php  if(isset($_SESSION['user_id'])){ ?>
+<th>Action</th><tr> <?php }?>
       <?php
            while($result = $stmt->fetch()){
                ?>
                <tr>
                     <td><a href="view.php?profile_id= <?php echo $result["profile_id"] ?>"><?php echo $result['first_name']; ?></a></td>
                     <td><?php echo $result['headline']; ?></td>
-                    <td><a href='edit.php?profile_id= <?php echo $result["profile_id"] ?>'> Edit</a>
+                    
+                    <?php  if(isset($_SESSION['user_id'])){
+
+                        ?><td><a href='edit.php?profile_id= <?php echo $result["profile_id"] ?>'> Edit</a>
                         <a href="delete.php?profile_id= <?php echo $result["profile_id"] ?>"> Delete</a>
-                    </td>
+                        </td>
+                        <?php
+
+                    }?>
+                  
                </tr>
                <?php
